+++ djangojs/__init__.py
@@ -7,4 +7,5 @@ __version__ = '0.6.dev'
 __description__ = "Django JS Tools"
 
 #: Packaged jQuery version
-JQUERY_VERSION = '1.8.3'
+JQUERY_DEFAULT_VERSION = '1.9.0'
+JQUERY_MIGRATE_VERSION = '1.0.0'